requirements.txt
numpy
numba
easydict
pyyaml
scikit-image
tqdm
kornia
torchvision
